requirements.txt
Django==4.1.5
djangorestframework==3.14.0
python-dotenv==0.14.0
psycopg2-binary==2.8.5
requests==2.28.1
celery[pytest,redis]==5.2.7
django-redis==5.2.0
redis==3.5.3
django-environ==0.4.5
google-cloud-secret-manager==2.6.0
django-storages[google]==1.11.1
pandas==1.5.3
algoliasearch-django==2.0
gunicorn>=20.1.0,<20.2
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.2.2
redis
